fixed error when loading only 1 match as next upcoming match

+++ app/Http/Controllers/pages/MatchController.php
@@ -17,6 +17,7 @@ class MatchController extends Controller
     public function remove($id)
     {
         $match = Match::find($id);
+        if(!$match) return \Redirect::back()->withErrors(['Wedstrijd niet gevonden.']);
         $match->delete();
         return redirect( route('tournaments') );
     }
@@ -27,74 +28,73 @@ class MatchController extends Controller
     public function edit(Request $request, $updating=true)
     {
         $this->validator($request->all())->validate();  
-        if(Auth::Check()) 
+        if(!Auth::Check())  return \Redirect::back()->withErrors(['Je moet ingelod zijn om een wedstrijd aan te passen.']);
+        $user = Auth::user();   
+        $role = Role::Find($user->role_id); 
+        if($role->permission < 50 && $role->permission !== 20)  return \Redirect::back()->withErrors(['Je hebt niet de juiste permissies om een wedstrijd aan te passen.']);
+        if($updating)   
         {
-            $user = Auth::user();   
-            $role = Role::Find($user->role_id); 
-            if($role->permission >= 50 || $role->permission == 20)  
-            {
-                if($updating)   
-                {
-                    $match = Match::find($request->id);
-                    $result1 = Result::find($match->result1->id);
-                    $result2 = Result::find($match->result2->id);
-                }
-                else    
-                {
-                    $team_controller = new TeamController;
-                    $match = new Match();
-                    $result1 = new Result();
-                    $result2 = new Result();
-                    $match->has_ended = 0;
-                    if($request->team1_id)
-                        $result1->team_id = $request->team1_id;
-                    else
-                    {
-                        $team = $team_controller->search($request->team1);
-                        $team_id = json_decode(json_encode($team))->original[0]->id;
-                        if($team) $result1->team_id = $team_id;
-                        else        abort(404); 
-                    }
-                    if($request->team2_id)
-                        $result2->team_id = $request->team2_id;
-                    else
-                    {
-                        $team = $team_controller->search($request->team2);
-                        $team_id = json_decode(json_encode($team))->original[0]->id;
-                        if($team) $result2->team_id = $team_id;
-                        else        abort(404); 
-                    }
-                    $result1->save();
-                    $result2->save();
-                    $match->result1_id = $result1->id;
-                    $match->result2_id = $result2->id;
-                    $match->save();
-                }
-                if($request->date)
-                {
-                    $time = ($request->time) ? $request->time : '';
-                    $start = date('Y-m-d H:i:s', strtotime("$request->date $request->time"));
-                    $match->start = $start;
-                }
-                if($request->field) $match->field = $request->field;
-                if($request->tournament_id) $match->tournament_id = $request->tournament_id;
-                if($request->score1 !== null && $request->score2 !== null)
-                {
-                    $result1->score = $request->score1;
-                    $result2->score = $request->score2;
-                    $match->has_ended = 1;
-                }
-                $result1->save();
-                $result2->save();
-                $match->save();
-                if(!$updating && $request->tournament_id)
-                    return redirect(route('match.create.route', $request->tournament_id));
-                elseif($match->id)
-                    return redirect(route('match.edit.route', $match->id));
-                else
-                    return redirect(route('tournaments'));
-            }
+            $match = Match::find($request->id);
+            if(!$match) return \Redirect::back()->withErrors(['Wedstrijd niet gevonden.']);
+            $result1 = Result::find($match->result1->id);
+            $result2 = Result::find($match->result2->id);
         }
+        else    
+        {
+            $team_controller = new TeamController;
+            $match = new Match();
+            $result1 = new Result();
+            $result2 = new Result();
+            $match->has_ended = 0;
+            if($request->team1_id)
+                $result1->team_id = $request->team1_id;
+            else
+            {
+                $team = $team_controller->search($request->team1);
+                $team_id = json_decode(json_encode($team))->original[0]->id;
+                if($team) $result1->team_id = $team_id;
+                else    return \Redirect::back()->withErrors(['Team 1 niet gevonden.']);
+            }
+            if($request->team2_id)
+                $result2->team_id = $request->team2_id;
+            else
+            {
+                $team = $team_controller->search($request->team2);
+                $team_id = json_decode(json_encode($team))->original[0]->id;
+                if($team) $result2->team_id = $team_id;
+                else    return \Redirect::back()->withErrors(['Team 2 niet gevonden.']);
+            }
+            $result1->save();
+            $result2->save();
+            $match->result1_id = $result1->id;
+            $match->result2_id = $result2->id;
+            $match->save();
+        }
+        if($request->date)
+        {
+            $time = ($request->time) ? $request->time : '';
+            $start = date('Y-m-d H:i:s', strtotime("$request->date $request->time"));
+            $match->start = $start;
+        }
+        if($request->field) $match->field = $request->field;
+        if($request->tournament_id) $match->tournament_id = $request->tournament_id;
+        if($request->score1 !== null && $request->score2 !== null)
+        {
+            $result1->score = $request->score1;
+            $result2->score = $request->score2;
+            $match->has_ended = 1;
+        }
+        $result1->save();
+        $result2->save();
+        $match->save();
+        if(!$updating && $request->tournament_id)
+            return redirect(route('match.create.route', $request->tournament_id));
+        elseif($request->tournament_id)
+            return redirect(route('tournament', $request->tournament_id));
+        elseif($match->id)
+            return redirect(route('match.edit.route', $match->id));
+        else
+            return redirect(route('tournaments'));
     }
     protected function validator(array $data)
     {
@@ -109,6 +109,7 @@ class MatchController extends Controller
     }
     public function showCreateForm($tournament_id)
     {
+        if(!Auth::check())  return \Redirect::back()->withErrors(['Je moet ingelogd zijn om een wedstrijd aan te maken.']);
         $user = Auth::user();
         $permission = $user->role()->first()->permission;
         return view("/pages/ud-match",
